move PHPUnitExtension and Subscriber into Extension namespace
