+++ src/TestSuite/Fixture/Extension/PHPUnitExtension.php
@@ -14,9 +14,8 @@ declare(strict_types=1);
  * @since         4.3.0
  * @license       https://opensource.org/licenses/mit-license.php MIT License
  */
-namespace Cake\TestSuite\Fixture;
+namespace Cake\TestSuite\Fixture\Extension;
 
-use Cake\TestSuite\Fixture\PHPUnitSubscriber\PHPUnitStartedSubscriber;
 use PHPUnit\Runner\Extension\Extension;
 use PHPUnit\Runner\Extension\Facade;
 use PHPUnit\Runner\Extension\ParameterCollection;
